Focal Engine was updated to the latest version; COLMAP data support was added to help view photogrammetry cameras, tie points, and original photos; AnalysisObject class now stores shifts that were applied to mesh or point cloud; A lot of debug code is still in main.cpp; Fixed some compiler warnings.

+++ SubSystems/ComplexityCore/Layers/DataLayer.cpp
@@ -428,7 +428,7 @@ void LayerInterpolationData::SetInterpolationFactor(float NewValue)
 	InterpolationFactor = NewValue;
 }
 
-int LayerInterpolationData::GetLayerCount()
+size_t LayerInterpolationData::GetLayerCount()
 {
 	return UsedLayerIDs.size();
 }
